+++ modules/pytket-braket/tests/backend_test.py
@@ -20,7 +20,7 @@ from hypothesis import given, strategies
 import numpy as np
 import pytest
 from pytket.extensions.braket import BraketBackend
-from pytket.circuit import Circuit, OpType, Qubit  # type: ignore
+from pytket.circuit import Circuit, OpType, Qubit, Bit  # type: ignore
 from pytket.pauli import Pauli, QubitPauliString  # type: ignore
 from pytket.utils.expectations import (
     get_pauli_expectation_value,
@@ -37,14 +37,16 @@ REASON = "PYTKET_RUN_REMOTE_TESTS not set (requires configuration of AWS storage
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_simulator() -> None:
-    b = BraketBackend(
-        device_type="quantum-simulator",
-        provider="amazon",
-        device="sv1",
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [{"device_type": "quantum-simulator", "provider": "amazon", "device": "sv1"}],
+    indirect=True,
+)
+def test_simulator(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.supports_shots
     c = Circuit(2).H(0).CX(0, 1)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     n_shots = 100
     h0, h1 = b.process_circuits([c, c], n_shots)
@@ -62,6 +64,7 @@ def test_simulator() -> None:
 
     # Circuit with unused qubits
     c = Circuit(3).H(1).CX(1, 2)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, 1)
     res = b.get_result(h)
@@ -70,8 +73,13 @@ def test_simulator() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_dm_simulator() -> None:
-    b = BraketBackend(device_type="quantum-simulator", provider="amazon", device="dm1")
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [{"device_type": "quantum-simulator", "provider": "amazon", "device": "dm1"}],
+    indirect=True,
+)
+def test_dm_simulator(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.supports_density_matrix
     c = Circuit(2).H(0).SWAP(0, 1)
     cc = b.get_compiled_circuit(c)
@@ -84,14 +92,16 @@ def test_dm_simulator() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_tn1_simulator() -> None:
-    b = BraketBackend(
-        device_type="quantum-simulator",
-        provider="amazon",
-        device="tn1",
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [{"device_type": "quantum-simulator", "provider": "amazon", "device": "tn1"}],
+    indirect=True,
+)
+def test_tn1_simulator(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.supports_shots
     c = Circuit(2).H(0).CX(0, 1)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     n_shots = 100
     h0, h1 = b.process_circuits([c, c], n_shots)
@@ -104,6 +114,7 @@ def test_tn1_simulator() -> None:
     assert sum(counts.values()) == n_shots
     # Circuit with unused qubits
     c = Circuit(3).H(1).CX(1, 2)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, 1)
     res = b.get_result(h)
@@ -112,12 +123,13 @@ def test_tn1_simulator() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_ionq() -> None:
-    b = BraketBackend(
-        device_type="qpu",
-        provider="ionq",
-        device="ionQdevice",
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [{"device_type": "qpu", "provider": "ionq", "device": "ionQdevice"}],
+    indirect=True,
+)
+def test_ionq(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.persistent_handles
     assert b.supports_shots
     assert not b.supports_state
@@ -162,10 +174,20 @@ def test_ionq() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_rigetti() -> None:
-    b = BraketBackend(
-        device_type="qpu", provider="rigetti", device="Aspen-M-1", region="us-west-1"
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [
+        {
+            "device_type": "qpu",
+            "provider": "rigetti",
+            "device": "Aspen-M-1",
+            "region": "us-west-1",
+        }
+    ],
+    indirect=True,
+)
+def test_rigetti(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.persistent_handles
     assert b.supports_shots
     assert not b.supports_state
@@ -196,11 +218,21 @@ def test_rigetti() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_rigetti_with_rerouting() -> None:
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [
+        {
+            "device_type": "qpu",
+            "provider": "rigetti",
+            "device": "Aspen-M-1",
+            "region": "us-west-1",
+        }
+    ],
+    indirect=True,
+)
+def test_rigetti_with_rerouting(authenticated_braket_backend: BraketBackend) -> None:
     # A circuit that requires rerouting to a non-fully-connected architecture
-    b = BraketBackend(
-        device_type="qpu", provider="rigetti", device="Aspen-M-1", region="us-west-1"
-    )
+    b = authenticated_braket_backend
     c = Circuit(4).CX(0, 1).CX(0, 2).CX(0, 3).CX(1, 2).CX(1, 3).CX(2, 3)
     c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, 10)
@@ -208,10 +240,20 @@ def test_rigetti_with_rerouting() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_oqc() -> None:
-    b = BraketBackend(
-        device_type="qpu", provider="oqc", device="Lucy", region="eu-west-2"
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [
+        {
+            "device_type": "qpu",
+            "provider": "oqc",
+            "device": "Lucy",
+            "region": "eu-west-2",
+        }
+    ],
+    indirect=True,
+)
+def test_oqc(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.persistent_handles
     assert b.supports_shots
     assert not b.supports_state
@@ -246,6 +288,7 @@ def test_local_simulator() -> None:
     assert b.supports_shots
     assert b.supports_counts
     c = Circuit(2).H(0).CX(0, 1)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     n_shots = 100
     h = b.process_circuit(c, n_shots)
@@ -352,6 +395,7 @@ def test_probabilities() -> None:
 def test_probabilities_with_shots() -> None:
     b = BraketBackend(local=True)
     c = Circuit(2).V(1).CX(1, 0).S(1)
+    c.measure_all()
     probs_all = b.get_probabilities(c, n_shots=10)
     assert len(probs_all) == 4
     assert sum(probs_all) == pytest.approx(1)
@@ -406,6 +450,7 @@ def test_default_pass() -> None:
 def test_shots_bits_edgecases(n_shots, n_bits) -> None:
     braket_backend = BraketBackend(local=True)
     c = Circuit(n_bits, n_bits)
+    c.measure_all()
 
     # TODO TKET-813 add more shot based backends and move to integration tests
     h = braket_backend.process_circuit(c, n_shots)
@@ -427,17 +472,19 @@ def test_shots_bits_edgecases(n_shots, n_bits) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_postprocess_ionq() -> None:
-    b = BraketBackend(
-        device_type="qpu",
-        provider="ionq",
-        device="ionQdevice",
-    )
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [{"device_type": "qpu", "provider": "ionq", "device": "ionQdevice"}],
+    indirect=True,
+)
+def test_postprocess_ionq(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
     assert b.supports_contextual_optimisation
     c = Circuit(2).H(0).CX(0, 1).Y(0)
+    c.measure_all()
     c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, n_shots=10, postprocess=True)
-    ppcirc = Circuit.from_dict(json.loads(cast(str, h[4])))
+    ppcirc = Circuit.from_dict(json.loads(cast(str, h[5])))
     ppcmds = ppcirc.get_commands()
     assert len(ppcmds) > 0
     assert all(ppcmd.op.type == OpType.ClassicalTransform for ppcmd in ppcmds)
@@ -445,9 +492,80 @@ def test_postprocess_ionq() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_retrieve_available_devices() -> None:
-    backend_infos = BraketBackend.available_devices()
+@pytest.mark.parametrize("authenticated_braket_backend", [None], indirect=True)
+def test_retrieve_available_devices(
+    authenticated_braket_backend: BraketBackend,
+) -> None:
+    backend_infos = authenticated_braket_backend.available_devices(
+        aws_session=authenticated_braket_backend._aws_session
+    )
     assert len(backend_infos) > 0
     # Test annealers are filtered out.
-    backend_infos = BraketBackend.available_devices(region="us-west-2")
+    backend_infos = authenticated_braket_backend.available_devices(
+        region="us-west-2", aws_session=authenticated_braket_backend._aws_session
+    )
     assert len(backend_infos) > 0
+
+
+def test_partial_measurement() -> None:
+    b = BraketBackend(local=True)
+    c = Circuit(4, 4)
+    c.H(0).CX(0, 1)
+    c.Measure(0, 1)
+    c.Measure(2, 0)
+    c = b.get_compiled_circuit(c)
+    n_shots = 100
+    h = b.process_circuit(c, n_shots)
+    res = b.get_result(h)
+    readouts = res.get_shots()
+    assert all(len(readouts[i]) == 2 for i in range(n_shots))
+    assert all(readouts[i][0] == 0 for i in range(n_shots))
+    counts = res.get_counts()
+    assert sum(counts.values()) == n_shots
+
+
+def test_multiple_indices() -> None:
+    b = BraketBackend(local=True)
+    c = Circuit(0, 2)
+    q0 = Qubit("Z", [0, 0])
+    q1 = Qubit("Z", [0, 1])
+    c.add_qubit(q0)
+    c.add_qubit(q1)
+    c.H(q0)
+    c.CX(q0, q1)
+    c.Measure(q0, Bit(0))
+    c.Measure(q1, Bit(1))
+    c1 = b.get_compiled_circuit(c)
+    h = b.process_circuit(c1, 100)
+    res = b.get_result(h)
+    readouts = res.get_shots()
+    assert all(readout[0] == readout[1] for readout in readouts)
+
+
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+@pytest.mark.parametrize(
+    "authenticated_braket_backend",
+    [
+        {
+            "device_type": "qpu",
+            "provider": "rigetti",
+            "device": "Aspen-M-1",
+            "region": "us-west-1",
+        }
+    ],
+    indirect=True,
+)
+def test_multiple_indices_rigetti(authenticated_braket_backend: BraketBackend) -> None:
+    b = authenticated_braket_backend
+    c = Circuit(0, 2)
+    q0 = Qubit("Z", [0, 0])
+    q1 = Qubit("Z", [0, 1])
+    c.add_qubit(q0)
+    c.add_qubit(q1)
+    c.H(q0)
+    c.CX(q0, q1)
+    c.Measure(q0, Bit(0))
+    c.Measure(q1, Bit(1))
+    c1 = b.get_compiled_circuit(c)
+    h = b.process_circuit(c1, 100)
+    b.cancel(h)