+++ modules/pytket-ionq/tests/backend_test.py
@@ -31,10 +31,14 @@ REASON = "PYTKET_RUN_REMOTE_TESTS not set (requires configuration of IoNQ API ke
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_small_circuit_ionq() -> None:
-    backend = IonQBackend(device_name="simulator", label="test 1")
-    # backend = IonQBackend("invalid", device_name="simulator", label="test 5")
-    # backend._MACHINE_DEBUG = True
+@pytest.mark.parametrize(
+    "authenticated_ionq_backend", [{"label": "test 1"}], indirect=True
+)
+def test_small_circuit_ionq(
+    authenticated_ionq_backend: IonQBackend,
+) -> None:
+    backend = authenticated_ionq_backend
+    # backend = IonQBackend(device_name="simulator", label="test 1")
 
     qc = Circuit(3, 3)
     qc.H(0)
@@ -44,16 +48,17 @@ def test_small_circuit_ionq() -> None:
     qc.Measure(2, 2)
     qc = backend.get_compiled_circuit(qc)
     counts = backend.run_circuit(qc, n_shots=1000).get_counts()
-    # note that we are rebuilding counts from probabilities, which are
-    # floats, and therefore count number is not always preserved!
-    assert counts[(0, 0, 0)] > 498 and counts[(0, 0, 0)] < 502
-    assert counts[(0, 1, 1)] > 498 and counts[(0, 1, 1)] < 502
     assert sum(counts.values()) == 1000
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_big_circuit_ionq() -> None:
-    backend = IonQBackend(device_name="simulator", label="test 2")
+@pytest.mark.parametrize(
+    "authenticated_ionq_backend", [{"label": "test 2"}], indirect=True
+)
+def test_big_circuit_ionq(
+    authenticated_ionq_backend: IonQBackend,
+) -> None:
+    backend = authenticated_ionq_backend
     circ = Circuit(4)
     circ.X(0).Y(0).Z(0).H(1).S(1).Sdg(1).H(1).T(2).Tdg(2).V(3).Vdg(3)
     circ.SWAP(0, 1)
@@ -66,18 +71,19 @@ def test_big_circuit_ionq() -> None:
 
 def test_invalid_token() -> None:
     token = "invalid"
-    b = IonQBackend(api_key=token, device_name="simulator", label="test 3")
-    c = Circuit(2, 2).H(0).CX(0, 1)
-    c.measure_all()
-    c = b.get_compiled_circuit(c)
     with pytest.raises(RuntimeError) as excinfo:
-        b.process_circuits([c], 1)
-        assert "Invalid key provided" in str(excinfo.value)
+        IonQBackend(api_key=token, device_name="simulator", label="test 3")
+        assert "Unauthorized" in str(excinfo.value)
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_invalid_request() -> None:
-    b = IonQBackend(device_name="simulator", label="test 4")
+@pytest.mark.parametrize(
+    "authenticated_ionq_backend", [{"label": "test 4"}], indirect=True
+)
+def test_invalid_request(
+    authenticated_ionq_backend: IonQBackend,
+) -> None:
+    b = authenticated_ionq_backend
     c = Circuit(2, 2).H(0).CZ(0, 1)
     c.measure_all()
     with pytest.raises(CircuitNotValidError) as excinfo:
@@ -86,8 +92,9 @@ def test_invalid_request() -> None:
 
 
 def test_machine_debug() -> None:
-    b = IonQBackend(api_key="invalid", device_name="simulator", label="test 5")
-    b._MACHINE_DEBUG = True
+    b = IonQBackend(
+        api_key="invalid", device_name="simulator", label="test 5", _machine_debug=True
+    )
     c = Circuit(2)
     c.H(0)
     c.CX(0, 1)
@@ -98,8 +105,13 @@ def test_machine_debug() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_cancellation() -> None:
-    b = IonQBackend(device_name="simulator", label="test 6")
+@pytest.mark.parametrize(
+    "authenticated_ionq_backend", [{"label": "test 6"}], indirect=True
+)
+def test_cancellation(
+    authenticated_ionq_backend: IonQBackend,
+) -> None:
+    b = authenticated_ionq_backend
 
     qc = Circuit(3, 3)
     qc.H(0)
@@ -113,7 +125,7 @@ def test_cancellation() -> None:
 
 
 def test_default_pass() -> None:
-    b = IonQBackend(api_key="invalid", device_name="simulator")
+    b = IonQBackend(api_key="invalid", device_name="simulator", _machine_debug=True)
     for ol in range(3):
         comp_pass = b.default_compilation_pass(ol)
         c = Circuit(3, 3)
@@ -128,8 +140,13 @@ def test_default_pass() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_postprocess() -> None:
-    b = IonQBackend(device_name="simulator", label="test 7")
+@pytest.mark.parametrize(
+    "authenticated_ionq_backend", [{"label": "test 7"}], indirect=True
+)
+def test_postprocess(
+    authenticated_ionq_backend: IonQBackend,
+) -> None:
+    b = authenticated_ionq_backend
     assert b.supports_contextual_optimisation
     c = Circuit(2, 2)
     c.Rx(0.1, 0)
@@ -153,9 +170,8 @@ def test_postprocess() -> None:
 )
 def test_shots_bits_edgecases(n_shots, n_bits) -> None:
     ionq_backend = IonQBackend(
-        api_key="invalid", device_name="simulator", label="test 5"
+        api_key="invalid", device_name="simulator", label="test 5", _machine_debug=True
     )
-    ionq_backend._MACHINE_DEBUG = True
     c = Circuit(n_bits, n_bits)
     # TODO TKET-813 add more shot based backends and move to integration tests
     h = ionq_backend.process_circuit(c, n_shots)
@@ -177,8 +193,9 @@ def test_shots_bits_edgecases(n_shots, n_bits) -> None:
 
 
 def test_nshots_ionq() -> None:
-    b = IonQBackend(api_key="invalid", device_name="simulator", label="test 8")
-    b._MACHINE_DEBUG = True
+    b = IonQBackend(
+        api_key="invalid", device_name="simulator", label="test 8", _machine_debug=True
+    )
     circuit = Circuit(1, 1).Measure(0, 0)
     n_shots = [1, 2, 3]
     handles = b.process_circuits([circuit] * 3, n_shots=n_shots)
@@ -189,7 +206,9 @@ def test_nshots_ionq() -> None:
     assert [get_nshots(h) for h in handles] == n_shots
 
 
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_retrieve_available_devices() -> None:
-    backend_infos = IonQBackend.available_devices()
-    assert len(backend_infos) == 1
-    assert backend_infos[0].device_name == "qpu"
+    backend_infos = IonQBackend.available_devices(
+        api_key=os.getenv("PYTKET_REMOTE_IONQ_API_KEY")
+    )
+    assert len(backend_infos) > 0