+++ modules/pytket-quantinuum/tests/backend_test.py
@@ -14,7 +14,6 @@
 
 from collections import Counter
 from typing import cast, Callable, Any  # pylint: disable=unused-import
-from ast import literal_eval
 import json
 import os
 from hypothesis import given, settings, strategies
@@ -22,6 +21,7 @@ import numpy as np
 import pytest
 import hypothesis.strategies as st
 from hypothesis.strategies._internal import SearchStrategy
+from hypothesis import HealthCheck
 
 from pytket.circuit import (  # type: ignore
     Circuit,
@@ -38,6 +38,7 @@ from pytket.circuit import (  # type: ignore
 )
 from pytket.extensions.quantinuum import QuantinuumBackend
 from pytket.extensions.quantinuum.backends.quantinuum import (
+    DEVICE_FAMILY,
     GetResultFailed,
     _GATE_SET,
 )
@@ -53,9 +54,17 @@ REASON = (
     "PYTKET_RUN_REMOTE_TESTS not set (requires configuration of Quantinuum username)"
 )
 
+ALL_DEVICE_NAMES = ["H1-1SC", "H1-2SC", "H1", "H1-1", "H1-2", "H1-1E", "H1-2E"]
 
-def test_quantinuum() -> None:
-    backend = QuantinuumBackend(device_name="H1-1SC", machine_debug=skip_remote_tests)
+
+@pytest.mark.parametrize("authenticated_quum_backend", [None], indirect=True)
+def test_quantinuum(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    if skip_remote_tests:
+        backend = QuantinuumBackend(device_name="H1-1SC", machine_debug=True)
+    else:
+        backend = authenticated_quum_backend
     c = Circuit(4, 4, "test 1")
     c.H(0)
     c.CX(0, 1)
@@ -87,8 +96,13 @@ def test_quantinuum() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_bell() -> None:
-    b = QuantinuumBackend(device_name="H1-1SC")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_bell(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     c = Circuit(2, 2, "test 2")
     c.H(0)
     c.CX(0, 1)
@@ -101,8 +115,15 @@ def test_bell() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_multireg() -> None:
-    b = QuantinuumBackend(device_name="H1-1SC", label="test 3")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend",
+    [{"device_name": "H1-1SC", "label": "test 3"}],
+    indirect=True,
+)
+def test_multireg(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     c = Circuit()
     q1 = Qubit("q1", 0)
     q2 = Qubit("q2", 0)
@@ -124,8 +145,13 @@ def test_multireg() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_default_pass() -> None:
-    b = QuantinuumBackend(device_name="H1-1SC")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_default_pass(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     for ol in range(3):
         comp_pass = b.default_compilation_pass(ol)
         c = Circuit(3, 3)
@@ -140,8 +166,15 @@ def test_default_pass() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_cancel() -> None:
-    b = QuantinuumBackend(device_name="H1-1SC", label="test cancel")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend",
+    [{"device_name": "H1-1SC", "label": "test cancel"}],
+    indirect=True,
+)
+def test_cancel(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     c = Circuit(2, 2).H(0).CX(0, 1).measure_all()
     c = b.get_compiled_circuit(c)
     handle = b.process_circuit(c, 10)
@@ -191,26 +224,47 @@ def circuits(
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-@given(
-    circuits(),  # pylint: disable=no-value-for-parameter
-    st.integers(min_value=1, max_value=10000),
+@pytest.mark.parametrize(
+    "authenticated_quum_backend",
+    [{"device_name": name for name in ALL_DEVICE_NAMES}],
+    indirect=["authenticated_quum_backend"],
 )
-@settings(max_examples=5, deadline=None)
-def test_cost_estimate(c: Circuit, n_shots: int) -> None:
-
-    b = QuantinuumBackend("H1-1SC")
+@given(
+    c=circuits(),  # pylint: disable=no-value-for-parameter
+    n_shots=st.integers(min_value=1, max_value=10000),
+)
+@settings(
+    max_examples=5,
+    deadline=None,
+    suppress_health_check=[HealthCheck.function_scoped_fixture],
+)
+def test_cost_estimate(
+    authenticated_quum_backend: QuantinuumBackend,
+    c: Circuit,
+    n_shots: int,
+) -> None:
+    b = authenticated_quum_backend
     c = b.get_compiled_circuit(c)
-    estimate = b.cost_estimate(c, n_shots)
-    status = b.circuit_status(b.process_circuit(c, n_shots))
-    status_msg = status.message
-    message_dict = literal_eval(status_msg)
-    if message_dict["cost"] is not None:
-        api_cost = float(message_dict["cost"])
-        assert estimate == api_cost
+    if b._device_name == DEVICE_FAMILY:
+        with pytest.raises(ValueError) as e:
+            _ = b.cost(c, n_shots)
+        assert "Cannot find syntax checker" in str(e.value)
+        estimate = b.cost(c, n_shots, syntax_checker="H1-1SC")
+    else:
+        estimate = b.cost(c, n_shots)
+    if estimate is None:
+        pytest.skip("API is flaky, sometimes returns None unexpectedly.")
+    assert isinstance(estimate, float)
+    assert estimate > 0.0
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_classical() -> None:
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_classical(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
     # circuit to cover capabilities covered in example notebook
     c = Circuit(1)
     a = c.add_c_register("a", 8)
@@ -234,15 +288,20 @@ def test_classical() -> None:
     c.X(0, condition=reg_geq(a, 1))
     c.X(0, condition=reg_leq(a, 1))
 
-    b = QuantinuumBackend("H1-1SC")
+    b = authenticated_quum_backend
 
     c = b.get_compiled_circuit(c)
     assert b.run_circuit(c, n_shots=10).get_counts()
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_postprocess() -> None:
-    b = QuantinuumBackend("H1-1SC")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_postprocess(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     assert b.supports_contextual_optimisation
     c = Circuit(2, 2)
     c.add_gate(OpType.PhasedX, [1, 1], [0])
@@ -289,10 +348,15 @@ def test_shots_bits_edgecases(n_shots, n_bits) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_simulator() -> None:
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1E"}], indirect=True
+)
+def test_simulator(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
     circ = Circuit(2, name="sim_test").H(0).CX(0, 1).measure_all()
     n_shots = 1000
-    state_backend = QuantinuumBackend("H1-1E")
+    state_backend = authenticated_quum_backend
     stabilizer_backend = QuantinuumBackend("H1-1E", simulator="stabilizer")
 
     circ = state_backend.get_compiled_circuit(circ)
@@ -328,8 +392,11 @@ def test_simulator() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_retrieve_available_devices() -> None:
-    backend_infos = QuantinuumBackend.available_devices()
+@pytest.mark.parametrize("authenticated_quum_backend", [None], indirect=True)
+def test_retrieve_available_devices(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    backend_infos = authenticated_quum_backend.available_devices()
     assert len(backend_infos) > 0
     api_handler = QuantinuumAPI()
     backend_infos = QuantinuumBackend.available_devices(api_handler=api_handler)
@@ -337,9 +404,14 @@ def test_retrieve_available_devices() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_batching() -> None:
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1E"}], indirect=True
+)
+def test_batching(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
     circ = Circuit(2, name="batching_test").H(0).CX(0, 1).measure_all()
-    state_backend = QuantinuumBackend("H1-1E")
+    state_backend = authenticated_quum_backend
     circ = state_backend.get_compiled_circuit(circ)
 
     handles = state_backend.process_circuits([circ, circ], 10)
@@ -355,8 +427,13 @@ def test_batching() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_submission_with_group() -> None:
-    b = QuantinuumBackend(device_name="H1-1SC")
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_submission_with_group(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    b = authenticated_quum_backend
     c = Circuit(2, 2, "test 2")
     c.H(0)
     c.CX(0, 1)
@@ -369,8 +446,13 @@ def test_submission_with_group() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_zzphase() -> None:
-    backend = QuantinuumBackend(device_name="H1-1SC", machine_debug=skip_remote_tests)
+@pytest.mark.parametrize(
+    "authenticated_quum_backend", [{"device_name": "H1-1SC"}], indirect=True
+)
+def test_zzphase(
+    authenticated_quum_backend: QuantinuumBackend,
+) -> None:
+    backend = authenticated_quum_backend
     c = Circuit(2, 2, "test rzz")
     c.H(0)
     c.CX(0, 1)
@@ -392,3 +474,9 @@ def test_zzphase() -> None:
     res = backend.get_result(handle, timeout=49)
     counts = res.get_counts()
     assert counts == correct_counts
+
+
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+@pytest.mark.parametrize("device_name", ALL_DEVICE_NAMES)
+def test_device_state(device_name: str) -> None:
+    assert isinstance(QuantinuumBackend.device_state(device_name), str)