+++ modules/pytket-stim/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.9.0"
+__extension_version__ = "0.10.0"
 __extension_name__ = "pytket-stim"